// WebsiteClient/src/app/Core/Interfaces/Components/IComponent.ts
export interface IComponent {
    id?: number;
    componentId?: number;
    componentKey?: string;
    componentName?: string;
    languageId?: string;
    name?: string;
    description?: string;
    details?: string;
    tags?: string;
    title?: string;
    supTitle?: string;
    link?: string;
    photoUrl?: string;
    startDate?: Date;
    endDate?: Date;
    date?: Date;
    htmlContent?: string;
    active?: boolean;
    publish?: boolean;
    added?: boolean;
}